+++ src/main/java/net/geforcemods/securitycraft/ConfigHandler.java
@@ -53,7 +53,6 @@ public class ConfigHandler {
 	}
 
 	public static class Server {
-		public DoubleValue codebreakerChance;
 		public BooleanValue allowAdminTool;
 		public BooleanValue shouldSpawnFire;
 		public BooleanValue ableToBreakMines;
@@ -83,11 +82,6 @@ public class ConfigHandler {
 
 		Server(ModConfigSpec.Builder builder) {
 			//@formatter:off
-			codebreakerChance = builder
-					.comment("The chance for the codebreaker to successfully hack a block. 0.33 is 33%. Set to a negative value to disable the codebreaker.",
-							"Using the codebreaker when this is set to 0.0 will still damage the item, while negative values do not damage it.")
-					.defineInRange("codebreaker_chance", 0.33D, -1.0D, 1.0D);
-
 			allowAdminTool = builder
 					.comment("Can the admin tool be used?")
 					.define("allowAdminTool", true);